Add detailed comments to all code files
