--- .env
@@ -1,3 +1,0 @@
-SECRET_KEY=your_secret_key
-DATABASE_URL=your_database_url
-
