+++ backend/.env
@@ -1,0 +1,2 @@
+SECRET_KEY=your_secret_key
+DATABASE_URL=your_database_url
