+++ backend/app.py
@@ -1,0 +1,16 @@
+from flask import Flask
+from config import Config
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+
+app = Flask(__name__)
+app.config.from_object(Config)
+
+db = SQLAlchemy(app)
+migrate = Migrate(app, db)
+
+from routes import init_routes
+init_routes(app)
+
+if __name__ == '__main__':
+    app.run(debug=True)
