+++ backend/static/images/Progress-Tracking
@@ -1,0 +1,1 @@
+"C:\Users\Maxyn\Desktop\Progress tracking.gif"
