+++ backend/static/images/personalized_workout_plans
@@ -1,0 +1,1 @@
+"C:\Users\Maxyn\Desktop\customize-workout-plans.jpg"
