+++ backend/utils.py
@@ -1,0 +1,7 @@
+import hashlib
+
+def hash_password(password):
+    return hashlib.sha256(password.encode()).hexdigest()
+
+def verify_password(stored_password, provided_password):
+    return stored_password == hash_password(provided_password)
