--- config.py
@@ -1,8 +1,0 @@
-import os
-
-class Config:
-    SECRET_KEY = os.environ.get('SECRET_KEY') or 'a_super_secret_key'
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL') or \
-        'sqlite:///app.db'
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
