--- footer.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import './Footer.css';
-
-/**
- * Footer component.
- * Displays footer information for the application.
- */
-function Footer() {
-  return (
-    <footer className="Footer">
-      <p>&copy; 2024 MXVirtualFit. All rights reserved.</p>
-    </footer>
-  );
-}
-
-export default Footer;
-
