+++ frontend/public/logo192.png
@@ -1,0 +1,1 @@
+"C:\Users\Maxyn\Desktop\IMG_20240717_200525.jpg"
