+++ frontend/public/robots.txt
@@ -1,0 +1,5 @@
+# robots.txt
+
+User-agent: *
+Disallow:
+
