+++ frontend/src/components/script.js
@@ -1,0 +1,3 @@
+document.addEventListener('DOMContentLoaded', function() {
+    console.log('JavaScript loaded!');
+});
