--- requirement.txt
@@ -1,5 +1,0 @@
-Flask==2.0.1
-Flask-SQLAlchemy==2.5.1
-Flask-Migrate==3.1.0
-python-dotenv==0.19.1
-
