--- styles.css
@@ -1,14 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f8f9fa;
-    margin: 0;
-    padding: 0;
-}
-
-.header {
-    background-color: #343a40;
-    color: #fff;
-    padding: 10px 0;
-    text-align: center;
-}
-
